requirements.txt
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0
gunicorn==21.2.0
simple-websocket==1.1.0
python-dotenv==1.0.1
openai-whisper==20231117
moviepy==1.0.3
pydub==0.25.1
numpy==1.24.3